+++ system/libraries/Form_validation.php
@@ -1130,6 +1130,25 @@ class CI_Form_validation {
 	// --------------------------------------------------------------------
 
 	/**
+	 * Exists
+	 *
+	 * Check if the input value already exists
+	 * in the specified database field.
+	 *
+	 * @param	string	$str
+	 * @param	string	$field
+	 * @return	bool
+	 */
+	public function exists($str, $field)
+	{
+		sscanf($field, '%[^.].%[^.]', $table, $field);
+		return isset($this->CI->db)
+			? ($this->CI->db->get_where($table, array($field => $str))->num_rows() !== 0)
+			: FALSE;
+	}
+
+	// --------------------------------------------------------------------
+	/**
 	 * Minimum Length
 	 *
 	 * @param	string